+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name='korpus',
-    packages=['korpus'],
+    py_modules=['korpus'],
     version='0.0.1',
     description='similarity made easy!',
     author='Metglobal',
